Removed unused files and changes path
