+++ src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Login from './components/Login';
-import Dashboard from './components/dashboard/Dashboard';
-import Home from './components/Home';
-import SinglePost from './components/SinglePost';
-import CreatePost from './components/dashboard/posts/CreatePost';
+import Login from './pages/Login';
+import Dashboard from "../src/pages/Dashboard"
+import Home from '../src/pages/Home';
+import SinglePost from './pages/SinglePost';
+import CreatePost from './pages/CreatePost';
 import AppProvider from './components/context/AppProvider';
-import Posts from './components/dashboard/posts/Posts';
+import Posts from './pages/AllPost';
 
 class App extends React.Component {
   render() {
