+++ src/api/constants.js
@@ -1,0 +1,5 @@
+const baseUrl = {
+	siteUrl: 'https://www.joakimvanebo.info',
+};
+
+export default baseUrl;
