--- src/client-config.js
@@ -1,5 +1,0 @@
-const clientConfig = {
-	siteUrl: 'https://www.joakimvanebo.info',
-};
-
-export default clientConfig;
