+++ src/pages/AllPost.js
@@ -1,43 +1,40 @@
 import React, { useEffect, useState } from 'react';
-import Navbar from './Navbar';
-import Footer from "./Footer";
-import Loader from '../assets/loader/loader.gif';
-import { Link } from 'react-router-dom';
-import Moment from 'react-moment';
-import clientConfig from '../client-config';
-import './style/GlobalStyle.css';
-import './style/Navbar.css';
-import './style/Home.css';
-import './style/Footer.css';
+import axios from "axios";
+import { Link } from "react-router-dom";
+import Moment from "react-moment";
+import Api from "../api/constants";
+import Loader from "../assets/loader/loader.gif";
+import DeletePost from "./DeletePost";
+import Navbar from "../components/Navbar";
 
-const Home = () => {
+const Posts = () => {
   const [loading, setLoading] = useState(false);
   const [posts, setPosts] = useState([]);
-  const [error, setError] = useState('');
-  const [searchQuery, setSearchQuery] = useState('');
+  const [error, setError] = useState("");
+  const [searchQuery, setSearchQuery] = useState("");
 
   useEffect(() => {
     fetchPosts();
   }, []);
 
   const fetchPosts = () => {
-    const wordPressSiteURL = clientConfig.siteUrl;
+    const wordPressSiteURL = Api.siteUrl;
 
     setLoading(true);
-    setError('');
+    setError("");
 
-    fetch(`${wordPressSiteURL}/wp-json/wp/v2/posts/`)
+    axios
+      .get(`${wordPressSiteURL}/wp-json/wp/v2/posts/`)
       .then((res) => {
-        if (res.ok) {
-          return res.json();
-        } else {
-          throw new Error('No Posts Found');
+        if (res.status === 200) {
+          if (res.data.length) {
+            setPosts(res.data);
+          } else {
+            setError("No Posts Found");
+          }
         }
       })
-      .then((data) => {
-        setPosts(data);
-      })
-      .catch((err) => setError(err.message))
+      .catch((err) => setError(err))
       .finally(() => setLoading(false));
   };
 
@@ -46,8 +43,8 @@ const Home = () => {
   };
 
   const filterPosts = () => {
-    if (searchQuery.trim() === '') {
-      return posts;
+    if (searchQuery.trim() === "") {
+      return posts; // If search query is empty, return all posts
     }
 
     const filteredPosts = posts.filter((post) =>
@@ -63,14 +60,8 @@ const Home = () => {
     <>
       <Navbar />
       {error && (
-        <div
-          className="alert alert-danger"
-          dangerouslySetInnerHTML={{ __html: error }}
-        />
+        <div className="alert alert-danger" dangerouslySetInnerHTML={{ __html: error }} />
       )}
-      <div className="centered-heading">
-        <h1>Semester Project</h1>
-      </div>
       <div className="search-container mt-3">
         <input
           type="text"
@@ -85,12 +76,13 @@ const Home = () => {
             <div
               key={post.id}
               className="card border-dark mb-3"
-              style={{ maxWidth: '50rem' }}
+              style={{ maxWidth: "50rem" }}
             >
               <div className="card-header">
                 <Link
                   to={`/post/${post.id}`}
                   className="text-secondary font-weight-bold"
+                  style={{ textDecoration: "none" }}
                 >
                   {post.title.rendered}
                 </Link>
@@ -102,23 +94,25 @@ const Home = () => {
                 />
               </div>
               <div className="card-footer">
-                <Moment format="MMMM Do, YYYY">{post.date}</Moment>
+                <Moment fromNow>{post.date}</Moment>
                 <Link
                   to={`/post/${post.id}`}
-                  className="btn btn-secondary float-right read-more-btn"
-                  style={{ textDecoration: 'none', marginRight: '10px' }}
+                  className="btn btn-secondary float-right"
+                  style={{ textDecoration: "none" }}
                 >
                   Read More...
                 </Link>
+                <DeletePost postId={post.id} />
               </div>
             </div>
           ))}
         </div>
-      ) : null}
+      ) : (
+        ""
+      )}
       {loading && <img className="loader" src={Loader} alt="Loader" />}
-      <Footer />
     </>
   );
 };
 
-export default Home;
+export default Posts;
