+++ src/style/SinglePost.css
